engine: guard mid-run flow re-eval against chunk-ring exhaustion
Both PR #632 reviewers flagged a panic: the post-loop root-flow re-eval
added for #625 called borrow_two(curr_chunk, next_chunk) unconditionally,
but run_to(target) with target past FINAL_TIME exits the loop via the
chunk-ring exhaustion break in save_advance!, which sets
curr_chunk = next_chunk before breaking. borrow_two with two equal chunk
indices slices left[a*n_slots..(a+1)*n_slots] out of a left of length
a*n_slots and panics -- aborting across the FFI boundary on what was
previously a graceful clamp (e.g. an unclamped simlin_sim_run_to).

Guard the re-eval on curr_chunk != next_chunk. The exhausted-slab case is
exactly the one a mid-interval read never reaches (a full slab means time
has reached FINAL_TIME, not mid-interval), so skipping the re-eval there
is correct and restores the pre-#625 graceful behavior. Every normal
run_to (mid-interval included) breaks via the curr[TIME] > end path with
distinct chunks, so the re-eval still runs and #625 stays fixed.

Also correct the imprecise "idempotent for the full-run path" note: on the
VM a full run leaves curr on the next-written overshoot chunk (stale
flow/aux), so the re-eval does change it there -- but that chunk is beyond
the saved results, so it stays invisible to a full run. The wasm note is
tightened too (it does break with a consistent curr, and has no chunk
aliasing since curr is the fixed CURR_BASE region).

Adds a VM regression test (run_to 10x past the end returns Ok, not panic).
The new test pushed vm.rs over the 6000-line file lint, so the
vm_reset_and_run_initials_tests module is extracted to a sibling file
included via #[path] (the existing vm_set_value_tests.rs pattern), so
use super::* still resolves vm's private items.

The residual VM-vs-wasm divergence for run_to(past FINAL_TIME) (VM
overshoots the slab; wasm clamps) is tracked as #634 -- not reachable via
in-range scrubbing.
